// src/resources/regions/health.ts
// File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.

import { APIResource } from '../../core/resource';
import { APIPromise } from '../../core/api-promise';
import { buildHeaders } from '../../internal/headers';
import { RequestOptions } from '../../internal/request-options';

export class Health extends APIResource {
  check(options?: RequestOptions): APIPromise<void> {
    return this._client.get('/regions/health', {
      ...options,
      headers: buildHeaders([{ Accept: '*/*' }, options?.headers]),
    });
  }

  performCheck(options?: RequestOptions): APIPromise<void> {
    return this._client.post('/regions/health/check', {
      ...options,
      headers: buildHeaders([{ Accept: '*/*' }, options?.headers]),
    });
  }
}